Reorganized directory structure and updated imports to reflect updates
