+++ nomad/adapt/continuous_spawn.py
@@ -13,10 +13,10 @@ parent(r,p,st,ti) ------------------------------------> child(r,p',st',ti)
    enforce constraint that classical energies be equal.
 """
 import numpy as np
-import nomad.parse.glbl as glbl
-import nomad.parse.log as log
-import nomad.utils.timings as timings
-import nomad.grow.utilities as utilities
+import nomad.simulation.glbl as glbl
+import nomad.simulation.log as log
+import nomad.simulation.timings as timings
+import nomad.adapt.utilities as utilities
 
 
 @timings.timed
