+++ nomad/initconds/explicit.py
@@ -2,8 +2,8 @@
 Sample a specific geometry or set of geometries.
 """
 import numpy as np
-import nomad.parse.glbl as glbl
-import nomad.basis.trajectory as trajectory
+import nomad.simulation.glbl as glbl
+import nomad.simulation.trajectory as trajectory
 
 
 def set_initial_coords(master):
