+++ nomad/initconds/general.py
