+++ nomad/propagators/runge_kutta4.py
@@ -11,9 +11,9 @@ Routines for propagation with the 4th order Runge-Kutta algorithm.
   ky4 = f[t + dt, y(t) + ky3*dt]
 """
 import numpy as np
-import nomad.parse.glbl as glbl
-import nomad.utils.timings as timings
-import nomad.dynamics.evaluate as evaluate
+import nomad.simulation.glbl as glbl
+import nomad.simulation.timings as timings
+import nomad.simulation.evaluate as evaluate
 
 
 rk_ordr = 4
