+++ nomad/simulation/atom_lib.py
@@ -1,7 +1,7 @@
 """
 A library containing atom widths, atomic numbers, etc.
 """
-import nomad.utils.constants as constants
+import nomad.math.constants as constants
 
 
 atom_name  = ['H', 'D', 'T', 'He', 'Li', 'Be', 'B', 'C', 'N', 'O', 'F',
